requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
certifi==2015.9.6.2
chardet==2.3.0
dj-database-url==0.3.0
Django==1.8.4
django-auth-ldap==1.2.6
django-bootstrap-pagination==0.1.10
django-celery==3.1.16
django-celery-with-redis==2.5
django-crispy-forms==1.1.2
django-debug-toolbar==0.9.4
django-extensions==0.8
django-floppyforms==0.4.7
django-indexer==0.3.0
django-model-utils==1.1.0
django-pagination==1.0.7
django-paging==0.2.4
django-picklefield==0.3.2
django-redis==2.2
django-redis-sessions==0.3.1
django-reversion==1.9.3
django-selectable==0.4.1
django-taggit==0.17.1
django-templatetag-sugar==0.1
django-tinymce==2.0.5
django-uni-form==0.9.0
django-wsgiserver==0.8.0rc1
docutils==0.8.1
hiredis==0.1.4
honcho==0.6.6
HTSQL==2.3.3
HTSQL-PGSQL==2.3.0
kombu==3.0.26
Markdown==2.6.2
ordereddict==1.1
psycopg2==2.6.1
Pycco==0.3.0
Pygments==2.0.2
pyparsing==1.5.6
pystache==0.5.4
python-ldap==2.4.20
pytz==2015.6
PyYAML==3.11
raven==1.6.0
redis==2.10.3
requests==0.11.1
selenium==2.21.1
simplejson==3.8.0
smartypants==1.8.6
South==0.7.3
SQLAlchemy==0.7.3
tablib==0.9.11
wheel==0.24.0
